+++ ChangeLog.md
@@ -1,7 +1,14 @@
 # CHANGELOG ADVANCEDPRODUCTSEARCH FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
+## 1.1
+- FIX : Wrong parameter for multiprice - *02/02/2021* - 1.1.4
+- FIX : Page reload prevent action param in url or post - *19/01/2021* - 1.1.3
+- FIX : Missing langs and help infos - *13/12/2021* - 1.1.2
+- FIX : By default, no supplier price preselection on supplier proposal, and conf ADVANCED_PRODUCT_SEARCH_PRESELECT_IF_ONE_FOURN_PRICE_ON_SUPPLIER_PROPOSAL to restore preselection - *08/12/2021* - 1.1.1
+- NEW : handle supplier documents - *08/12/2021* - 1.1.0
+
 ## 1.0
 - FIX : DA024689 paramètres inversé dans la fonction getSellPrice de l'object product qui ne remonte pas les  les niveaux de prix du tiers affecté à la propal - *23/04/2024* - 1.0.3  
 - FIX : We can't add products or services after pop-in page change - *2022-01-20* - 1.0.2
-- FIX : signature for `addline` not consistent across objects (proposal, order, invoice) - *2021-10-08* - 1.0.1
+- FIX: signature for `addline` not consistent across objects (proposal, order, invoice) - *08/10/2021* - 1.0.1
 - Initial version